requirements.txt
lxml
Django
RPi.GPIO